+++ src/SoftConstraint.cpp
@@ -1,6 +1,6 @@
 #include "../includes/SoftConstraint.h"
 
-SoftConstraint::SoftConstraint(string name, int value) {
+SoftConstraint::SoftConstraint(string name, double value) {
 	this->name = name;
 	this->value = value;
 }
